+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Releases are available [on GitHub][2].
 * Updated to GHC-9.0.1
 * Updated integration test suite
 * Updated file parser test suite
+* Updated TNT parser to use IUPAC codes from `Data.Alphabet.IUPAC`
 * Added GitHub Actions workflows for CI
 * Corrected build error in benchmarking suite
 * Corrected defect in string alignment commutativity